🛠️ Refactor: Restructure into src/ and implement centralized error handling
Extracted `MetadataDumper` out of `main.ts` to enforce the Single
Responsibility Principle. Created `src/` directory and moved all
typescript files to conform to directory conventions. Implemented a
centralized error reporting system (`src/errorReporter.ts`) to eliminate
empty catch blocks and funnel all unexpected errors properly. Created
`mise.toml` to replace ad-hoc tooling scripts. Codified conventions in
`AGENTS.md`.
