+++ rollup.config.mjs
@@ -13,7 +13,7 @@ if you want to view the source visit the plugins github repository
 `;
 
 export default {
-  input: 'main.ts',
+  input: 'src/main.ts',
   output: {
     file: 'dist/main.js',
     sourcemap: 'inline',
